Add and use `container-runtime-linux start`.
- Part of #653.
- No need for `defaultCommand` since ContainerService
  startup will write a new launchd plist with the
  `start` subcommand included.
